Sort matches based on mru buffer
Any buffer that's been edited that matches the search string is sorted
before other matches.

This changes matcher to use the score-based sort even in the query
string is blank. That gets it to respect the mru score modifications on
a blank search. I couldn't see why sorting on alpha without a query
would make a difference except as a minor optimization.